+++ NEWS.rst
@@ -8,6 +8,7 @@ Version 0.8.6 - dev
    * CHANGE: create all missing tables of the TABLES sections for DXF R12
    * BUGFIX: entities on new layouts will be saved
    * NEW: Layout.paper_setup() and correct 'main' viewport for DXF R2000+
+   * NEW: Table() and MText() composite objects from dxfwrite available in the ezdxf.addons subpackage
 
 Version 0.8.5 - 2018-01-28